requirements.txt
amqp==1.4.6
anyjson==0.3.3
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
billiard==3.3.0.20
celery==3.1.18
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
decorator==4.4.2
Django==3.0.8
django-redis==4.12.1
django-redis-cache==3.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==2.1.0
drf-yasg==1.20.0
eventlet==0.24.1
flower==0.9.7
future==0.18.2
gevent==21.1.1
greenlet==1.0.0
humanize==3.2.0
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
kombu==3.0.30
MarkupSafe==1.1.1
monotonic==1.5
openapi-codec==1.3.2
packaging==20.8
parso==0.8.1
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.10
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pymongo==3.11.2
pyparsing==2.4.7
pypiwin32==223
pyshorteners==1.0.1
pytest==6.2.1
pytest-django==4.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2015.4
pywin32==300
redis==2.10.3
requests==2.7.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
short-url==1.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
windows-curses==2.2.0
xlwt==1.3.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0